refactor: move notification services to notifications package and update imports accordingly
