+++ chatbot/src/main/java/pt/estga/chatbot/services/AuthService.java
@@ -6,7 +6,11 @@ import pt.estga.shared.models.AppPrincipal;
 import java.util.Optional;
 
 public interface AuthService {
+
     boolean isAuthenticated(String platformUserId);
+
     Optional<AppPrincipal> authenticate(String platformUserId);
+
     boolean supports(Platform platform);
+
 }
